# Inline and specialize Stdlib.{min,max}