--- a/stdlib/stdlib.ml
+++ b/stdlib/stdlib.ml
@@ -71,8 +71,8 @@ external ( <= ) : 'a -> 'a -> bool = "%lessequal"
 external ( >= ) : 'a -> 'a -> bool = "%greaterequal"
 external compare : 'a -> 'a -> int = "%compare"
 
-let min x y = if x <= y then x else y
-let max x y = if x >= y then x else y
+external min : 'a -> 'a -> 'a = "%min"
+external max : 'a -> 'a -> 'a = "%max"
 
 external ( == ) : 'a -> 'a -> bool = "%eq"
 external ( != ) : 'a -> 'a -> bool = "%noteq"