--- a/testsuite/tests/asmcomp/compare.ml
+++ b/testsuite/tests/asmcomp/compare.ml
@@ -8,3 +8,7 @@ let _ = f ()
 let[@inline never] myint () = print_string "bye\n"; 42
 let[@inline never] g () = compare (myint ()) 5;;
 let _ = g ()
+
+
+let () = Printf.printf "%.02f\n%!" (min (print_endline "world"; 43.) (float ()))
+let () = Printf.printf "%.02f\n%!" (max (print_endline "world"; 43.) (float ()))