--- a/testsuite/tests/asmcomp/compare.reference
+++ b/testsuite/tests/asmcomp/compare.reference
@@ -1,2 +1,8 @@
 hello
 bye
+hello
+world
+42.00
+hello
+world
+43.00