Fix #5541, fix #4808.

`Stdlib.{min,max}` are currently not inlined; this is because their size is above the default inlining threshold.  Compiling Stdlib with `-inline 2` would be enough to have them inlined, but since type-specialization of primitives happen before inlining, one ends up with a C call to the generic comparison, even if arguments are integers or floats.

This PR implements `Stdlib.{min,max}` with two new %-primitives `%min`, `%max`, applies the same specialization as for comparison primitives, then expands them to their definition when generating the Lambda code.  (No new runtime primitive has been added, so no need to adapt js_of_ocaml for instance.)

Some micro-bencharks:
````ocaml
(* int64 *)
let r = ref 0L in for i = 1 to 1_000_000_000 do r := max !r (Int64.of_int i) done
(* float *)
let r = ref 0. in for i = 1 to 1_000_000_000 do r := max !r (float i) done
(* int *)
let r = ref 0 in for i = 1 to 1_000_000_000 do r := max !r i done
(* string *)
let r = ref "" in for i = 1 to 1_000_000_000 do r := max !r (if i mod 2 = 1 then "Hello" else "World") done
````

|   | trunk | PR |  speedup
| ------------- | ------------- |------------- |------------- |
| int64 | 15s |  0.875s  |  x17 |
| float | 13.6s |  0.98s |  x13 |
| int | 6.0s | 0.48s |  x12 |
| string | 16.6s |  5.1s | x3 |

No noticeable impact for non-specialized cases.

Notes for the reviewer:

  - The first two commits are general cleanup on the primitive specialization code; they potentially lose some sharing of primitive descriptions in lambda code, but I suspect this is completely negligible.
  - The real stuff are the third and fifth commits, with a bootstrap of the compiler in-between.
  - There is no explicit performance test (I haven't seen any other test doing explicit time measurement).


Also, one could consider further improvements:

  - Better code generation in the backend (a branchless implementation for integers, for instance).
  - Explicit type specialized versions `{Int,Float,...}.{min,max}`