@alainfrisch 2021-03-08
@xavierleroy : since you were the one suspending #5541 do you have an opinion on this PR? (*"I'm not expecting a solution soon, which is why I'm putting this PR in suspended state."*  => indeed, that was 9 years ago!)