@xavierleroy 2021-03-08
#5541 was 9 years ago...  What I meant at the time is that type specialization after inlining the (OCaml) definition of min and max was far away in the future, and still is (?).  You sidestep the problem by turning min and max into primitives.  On the one hand it does the job.  On the other hand, it feels sad that operations that are perfectly definable in OCaml are being turned into primitives just to get better optimization.