@alainfrisch 2021-03-08
I perfectly agree, and I'll be happier with an inliner that would having to do that. But it also feels sad to know that the polymorphic comparison will be involved when min/max are used on simple numbers.  Do you think the workaround is decent enough?  If the inliner ever gets better, we can always get rid of it.  Or perhaps one should just expose specialized functions `{Int,Int32,Int64,Float}.{min,max}`  (and deprecate the functions in Stdlib?).

(One could argue that `not` is also definable in OCaml!  And all comparison could be derived from the single `%compare` operation -- except that's not really the case, but oh well.)