@alainfrisch 2021-04-28
@xavierleroy : what do you think we should do here?  Merge this PR, expose explicit functions like `Int.{min,max}` (also in Float, Int64, etc), accept the bad performance when people use min/max on numbers until we get a better inliner?