@xavierleroy 2021-04-29
I'm still not enthusiastic about adding "min" and "max" as primitives with specific rules for type specialization.

Half of the dev team wants to kill polymorphic equality.  While I stand up for polymorphic equality, I agree that polymorphic "min" and "max" are questionable and it would make sense to remove them.  Adding special compiler support for them goes against this grain.

Having specialized min and max functions in `Float`, `Int`, `Int32`, etc, would make more sense to me.  Plus, for `Float.min` and `Float.max` we'll get to argue about their semantics w.r.t. NaN.