@dbuenzli 2021-04-29
> Plus, for `Float.min` and `Float.max` we'll get to argue about their semantics w.r.t. NaN.

That [already happened](https://github.com/ocaml/ocaml/blob/9818f7aaa2a64a3b3d99ab73dd02caa5645ab9c0/stdlib/float.mli#L358-L368), the discussion starts [here](https://github.com/ocaml/ocaml/pull/1794#issuecomment-430274338).