@dbuenzli 2021-04-30
> (*) Even though it doesn't map to any "fmin" or "fmax" processor instruction I known of. so don't expect a super fast implementation any time soon.

Note that this PR also introduced the equivalent to the C99 `fmin` and `fmax` (NaN as missing data) under the names [`Float.{min,max}_num`](https://github.com/ocaml/ocaml/blob/9818f7aaa2a64a3b3d99ab73dd02caa5645ab9c0/stdlib/float.mli#L375-L385).