@alainfrisch 2021-05-03
> Having specialized min and max functions in Float, Int, Int32, etc, would make more sense to me.

Ok, this makes sense!  I've opened https://github.com/ocaml/ocaml/issues/10389 accordingly.

The first two commits are general code cleanup (which was useful to implement this PR), but I don't think it's worth spending reviewing time on them, so I'm closing this PR.